+++ src/Symfony/Component/Config/Builder/ClassBuilder.php
@@ -31,6 +31,7 @@ class ClassBuilder
     private array $require = [];
     private array $use = [];
     private array $implements = [];
+    private bool $allowExtraKeys = false;
 
     public function __construct(string $namespace, string $name)
     {
@@ -124,7 +125,7 @@ BODY
 
     public function addProperty(string $name, string $classType = null): Property
     {
-        $property = new Property($name, $this->camelCase($name));
+        $property = new Property($name, '_' !== $name[0] ? $this->camelCase($name) : $name);
         if (null !== $classType) {
             $property->setType($classType);
         }
@@ -159,5 +160,15 @@ BODY
     public function getFqcn(): string
     {
         return '\\'.$this->namespace.'\\'.$this->name;
+    }
+
+    public function setAllowExtraKeys(bool $allowExtraKeys): void
+    {
+        $this->allowExtraKeys = $allowExtraKeys;
+    }
+
+    public function shouldAllowExtraKeys(): bool
+    {
+        return $this->allowExtraKeys;
     }
 }